Updates config and env (patch)

+++ index.js
@@ -1,5 +1,6 @@
 const axios = require('axios')
 const path = require('path')
+const sleep = require('then-sleep')
 const addIndex = require('./lib/add-index')
 const deleteIndex = require('./lib/delete-index')
 const repackEmployee = require('./lib/repack-employee')
@@ -11,34 +12,38 @@ if (env) {
   const envFilePath = path.resolve(process.cwd(), env)
   logger('info', ['index', 'loading environment', env])
   require('dotenv').config({ path: envFilePath })
+} else if (process.env.NODE_ENV !== 'production') {
+  logger('info', ['index', 'loading environment', '.env'])
+  require('dotenv').config()
 } else {
   logger('warn', ['index', 'no environment loaded'])
 }
 
 async function indexAnsatte () {
   const { data } = await axios.get(process.env.SOURCE_URL)
-  const msg = await deleteIndex()
-  logger('info', ['index', 'indexAnsatte', 'index deleted', JSON.stringify(msg)])
+  if (!process.env.NO_DELETE_INDEX) {
+    const msg = await deleteIndex()
+    logger('info', ['index', 'indexAnsatte', 'index deleted', JSON.stringify(msg)])
+  } else {
+    logger('info', ['index', 'indexAnsatte', 'no delete index'])
+  }
   const indexes = data.results.map(repackEmployee).map(prepareForIndex)
   logger('info', ['index', 'index-ansatte', 'employees to index', indexes.length])
-  let jobs = indexes.map(addIndex)
-  let todo = indexes.length
-  let done = 0
   let success = 0
   let fail = 0
   const next = async () => {
-    if (done === todo) {
+    if (indexes.length === 0) {
       logger('info', ['index', 'indexAnsatte', 'finished', 'success', success, 'fail', fail])
     } else {
-      done++
-      const job = jobs.pop()
+      const index = indexes.pop()
       try {
-        await job
+        await addIndex(index)
         success++
       } catch (error) {
         fail++
         logger('error', ['index', 'indexAnsatte', error])
       }
+      sleep(500)
       await next()
     }
   }
